Added functionality to Camera
Camera can now be toggled on and off, displaying or not displaying it's data resectively. Memory can be erased when on. Additional changes (@seth if you see this before we meet, the code you pushed had a very different display than what you showed me. I hope that that was intentional or that I am crazy. Either way, camera has adequate functionality.)